+++ src/pushsource/_impl/backend/staged/staged_source.py
@@ -21,7 +21,6 @@ from .staged_utils import StagingMetadata, StagingLeafDir
 from .staged_ami import StagedAmiMixin
 from .staged_files import StagedFilesMixin
 from .staged_errata import StagedErrataMixin
-from .staged_channel_dumps import StagedChannelDumpsMixin
 from .staged_compsxml import StagedCompsXmlMixin
 from .staged_modulemd import StagedModuleMdMixin
 from .staged_productid import StagedProductIdMixin
@@ -38,7 +37,6 @@ class StagedSource(
     StagedAmiMixin,
     StagedFilesMixin,
     StagedErrataMixin,
-    StagedChannelDumpsMixin,
     StagedCompsXmlMixin,
     StagedModuleMdMixin,
     StagedProductIdMixin,