+++ src/pushsource/_impl/backend/staged/staged_unsupported.py
@@ -7,6 +7,7 @@ LOG = logging.getLogger("pushsource")
 
 class StagedUnsupportedMixin(StagedBaseMixin):
     @handles_type("DOCKER")
+    @handles_type("CHANNEL_DUMPS")
     def __push_item(self, _leafdir, _metadata, entry):
         LOG.error("Unsupported content found: %s", entry.path)
         return None