+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- n/a
+
+## [2.0.0] - 2020-11-04
+
 ### Added
+
+- Added some user-friendly logs when advisory or staged metadata fails validation.
 
 ### Removed
 
@@ -35,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial stable release of project
 
-[Unreleased]: https://github.com/release-engineering/pushsource/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/release-engineering/pushsource/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/release-engineering/pushsource/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/release-engineering/pushsource/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/release-engineering/pushsource/compare/v1.0.0...v1.1.0